backend/pyproject.toml
[project]
name = "fast-forward"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "alembic==1.17.2",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "beautifulsoup4==4.14.2",
    "black==25.11.0",
    "blinker==1.9.0",
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "click==8.3.0",
    "distro==1.9.0",
    "flask==3.1.2",
    "flask-cors==6.0.1",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "isort==7.0.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==1.1.0",
    "langchain-openai==1.1.0",
    "langsmith==0.4.47",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "mypy>=1.19.0",
    "mypy-extensions==1.1.0",
    "numpy>=2.3.5",
    "openai==2.8.1",
    "orjson==3.11.4",
    "packaging==25.0",
    "pathspec==0.12.1",
    "platformdirs==4.5.0",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "python-dotenv==1.2.1",
    "pytokens==0.3.0",
    "pyyaml==6.0.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "sqlalchemy==2.0.44",
    "tenacity==9.1.2",
    "tiktoken==0.12.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.5.0",
    "werkzeug==3.1.3",
    "zstandard==0.25.0",
]
